ai-news-bot/requirements.txt
feedparser==6.0.11
playwright==1.44.0
anthropic>=0.28.0
python-dotenv==1.0.1
